+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.1.2
+
+*Released: 2016-11-15*
+
+### Additions
+
+- Added `Lens.child(with:)` - @vknabel
+- Added `Lens.child(from:,to:)` - @vknabel
+- Added `Lens.asChild(of:)` - @vknabel
+- Added `Lens.asChild(from:,to:)` - @vknabel
+
 ## 0.1.1
 
 *Released: 2016-11-06*